Fix unpacking of particles in a VErtex
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@136951 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 7bbe0f9536fb8329a631d8f163b9d958a762603a

+++ LHCb/Event/PackedEvent/src/PackedVertex.cpp
@@ -120,7 +120,7 @@ void VertexPacker::unpack( const PackedData       & pvert,
     // outgoing particles
     for ( unsigned int iiP = pvert.firstOutgoingPart; iiP < pvert.lastOutgoingPart; ++iiP )
     {
-      const int & iP = pverts.outgoingParticles()[iiP];
+      const long long & iP = pverts.outgoingParticles()[iiP];
       int hintID(0), key(0);
       m_pack.hintAndKey64( iP, &pverts, &verts, hintID, key );
       SmartRef<LHCb::Particle> ref(&verts,hintID,key);
